ARQ-195 ARQ-216 Add client / container command based communication layer
Exposes a CommandService for Client and Container side to execute Commands and return results. Currently used by the Deployer API to communicate with the client side about triggering Deployments from whithin container.

Protocol has been extended to take getExecutor(Config, MetaData, CommandCallback). It's up to the Protocol to add the CommandService for incontainer, and handle the communication over the wire.

The Command itself is fired on the Client side and handled as a event. The Observer is responsible for setting the Result on the Command to return to Container side.