ARQ-605 Keep track of whether BeforeX was called so we can call AfterX
Add a local Stack to TestNG Arquillian integration to keep track of wheter our Before was called so we can call After if that is the case. TestNG configurations methods marked as alwaysRun will run regardless of groups or failure, we need to only call After when one of our Befores was called.

Note: One can use TestNG Groups to include all non Arquillian tests, but not to include Arquillian tests. Arquillians Configuration Methods does not belong to a group and can't inherit the subclasses groups.

Workaround: The Arquillian Configuration methods default belong to group "arquillian", include that group as well as your own.