ARQ-608 Register arquillian-testenricher-cdi as a CDI extension
This lets us obtain a BeanManager in non JNDI based environments like Weld-Servlet on Tomcat/Jetty. Original JNDI lookup is still preserved and attempted first. In a Modular deployment like EE, the BeanManager found by the CDI Extension might not be the BeanManager for the module we're requesting, so we rely on the application server to provide the correct JNDI context.