ARQ-1421 Check equality and not instance in BeansXMLProtocolProcessor
ARQ-891 changed sligtly the behavior of how the Archives are processed
in the Protocol Packagers. Instead of relying on casting of instances,
it now 'recreates' the Archive in the expected View via archive.as(Type).

That means applicationArchive != protocolArchive, but
applicationArchive.equals(protocolArchive) is still true.

Added more tests to BeansXMLProtocolProcessor.