ARQ-1975 Adds arquillian resource provider for servlet context in servlet protocol.
Adds ServletContext as Arquillian Resource so it can be used in the tests. In case of in-container tests, ServletContxt used is the one provided by the Servlet protocol.
In case of client test, it only works in case of embedded containers and they are the responsible of provide the instance by producing the instance